refactor: add Enums.cs for all flags in codebase
add the Enums.cs in the FPIS.Utils namespace to store all flags for the
various actions we might perform. This is extremely helpful in situations
where we might want to perform different actions based on the feedback
obtained from a function call. For example, during authentication we
might have to display a message when no employee id is provided. Or
when a password is not provided. Or when the authentication fails or
passes.