test: add unit test for Login.cs
add the LoginTest.cs unit test file with two test cases to easily determine
the functionality of the codes in the Login.cs source code file. Also
added public functions to easily test the functions in the Login.cs file. These
public functions would have to be deleted later on because they are only meant
for us and not for production.