fix: implement change request
implement the code changes requested by Patrick. Very similar to previous commit.
It turns out I forgot to save the recent changes in the editor so I winded up
commiting not up-to-date code. Made sure identifiers stick with a single naming convention.
Added blocks for the if statements I wrote in the Login.cs.