feat: track raw materials to be released
this feature is useful when samples of raw materials are
sent to Quality Control for analysis. Users can now track which
raw materials from a particular warehouse they would like to
issue after the Quality Control department provides their
results on the samples requested to be analyzed. A follow
up feature would enable the procurement department to specify
the quantity submitted for analysis